requirements.txt
pymongo >= 3.4.0
biopython >= 1.68
progressbar2 >= 3.11
findspark >= 1.1.0
xlrd >= 1.0.0
flask >= 0.12.1